+++ src/Areas/Kusto/Commands/TableSchemaCommand.cs
@@ -16,7 +16,10 @@ public sealed class TableSchemaCommand(ILogger<TableSchemaCommand> logger) : Bas
     public override string Name => "schema";
 
     public override string Description =>
-        "Get the schema of a specific table in an Kusto database.";
+        """
+        Get the schema of a specific table in an Kusto database.
+        Requires `cluster-uri` ( or `subscription` and `cluster-name`), `database-name` and `table`.
+        """;
 
     public override string Title => CommandTitle;